Flexible crediter that returns vectors and weights
The last one created a single vector using weights of the history that
generalized what occurred, but loses information on variation that
occurred over the period. This does not perform such a collapse, and
just returns a full set of weights and feature vectors, aligning more
closely with the Deep TAMER method.